+++ model/ICalculable.java
@@ -1,10 +1,10 @@
 package Seminar7.model;
 
 public interface ICalculable {
-    void sum(int x);
-    void multiplier(int x);
-    void subtruction(int x);
-    void division(int x);
-    void setFirstArg(int firstArg);
-    void getResult();
+    void sum(int x);      //   сумма
+    void multiplier(int x);  // умножение
+    void subtruction(int x);   // вычитание
+    void division(int x);   //   деление
+    void setFirstArg(double firstArg);  // установка 1 аргумента
+    void getResult();     // результат
 }
